Remove old captcha files and add new project inside app/
