--- requirements.txt
@@ -1,6 +1,0 @@
-tensorflow>=2.12.0
-numpy>=1.21.0
-pandas>=1.3.0
-opencv-python>=4.5.0
-Pillow>=8.0.0
-scikit-learn>=1.0.0
